requirements.txt
python-telegram-bot
flask
apscheduler
python-dotenv
requests